+++ packages/antd-ext/components/OutLineWrapper/OutLineWrapper.tsx
@@ -1,14 +1,24 @@
 import React, { useMemo } from 'react';
 import { Row, Col } from 'antd';
-import { InputExt } from './InputExt';
 import { css } from '@emotion/css';
-import type { InputProps } from 'antd';
 import { usePrefixCls } from '../utils';
+import type { GlobalToken } from 'antd';
 
-const style = (prefixCls: string) => css`
+export interface OutLineWrapperProps {
+  label: string;
+  children?: JSX.Element;
+  injectStyle?: (prefixCls: string, mtPrefixCls: string, token: GlobalToken) => string;
+}
+
+const style = (
+  prefixCls: string,
+  mtPrefixCls: string,
+  token: GlobalToken,
+  injectStyle?: (prefixCls: string, mtPrefixCls: string, token: GlobalToken) => string
+) => css`
   border: 1px solid #D0D3D6;
   border-radius: 4px;
-  padding: 0 13px;
+  padding-left: 13px;
 
   .${prefixCls}-form-item-label {
     height: 30px;
@@ -28,12 +38,6 @@ const style = (prefixCls: string) => css`
     }
   }
 
-  .${prefixCls}-input {
-    padding: 4px 0;
-    color: #333;
-    width: 100%;
-  }
-
   &:focus-within {
     border: 1px solid #2D64E5;
   }
@@ -41,24 +45,22 @@ const style = (prefixCls: string) => css`
   &:hover {
     border: 1px solid #477EFF;
   }
+
+  ${injectStyle ? injectStyle(prefixCls, mtPrefixCls, token) : ''}
 `;
 
-export interface InputOutLineExtProps extends InputProps {
-  label: string;
-}
-
-export default function InputOutLineExt(props: InputOutLineExtProps) {
-  const { label, bordered, ...restProps } = props;
-  const { prefixCls } = usePrefixCls()
-  const customClassName = useMemo(() => style(prefixCls), [prefixCls]);
+export default function OutLineWrapper<T extends OutLineWrapperProps>(props: T) {
+  const { label, children, injectStyle } = props;
+  const { token, prefixCls, mtPrefixCls } = usePrefixCls();
+  const customClassName = useMemo(() => style(prefixCls, mtPrefixCls, token, injectStyle), [prefixCls, mtPrefixCls, token]);
 
   return (
     <Row className={customClassName}>
       <Col className={`${prefixCls}-form-item-label`}>
-        <label htmlFor={restProps.id}>{label}</label>
+        <label>{label}</label>
       </Col>
       <Col className={`${prefixCls}-form-item-control`}>
-        <InputExt bordered={false} {...restProps} />
+        {children}
       </Col>
     </Row>
   );