+++ packages/antd-ext/components/RangeNumberExt/index.ts
@@ -1,0 +1,5 @@
+import RangeNumberExt from './RangeNumberExt';
+import RangeNumberOutLineExt from './RangeNumberOutLineExt';
+import { MixinRangeNumberExt, MixinRangeNumberOutLineExt } from './Mix';
+
+export { RangeNumberExt, RangeNumberOutLineExt, MixinRangeNumberExt, MixinRangeNumberOutLineExt };