+++ packages/antd-ext/components/index.ts
@@ -4,8 +4,10 @@ import TableExt from './TableExt';
 import { MixinInputExt as InputExt, MixinInputOutLineExt as InputOutLineExt } from './InputExt';
 import { MixinSelectExt as SelectExt, MixinSelectOutLineExt as SelectOutLineExt } from './SelectExt';
 import DrawerExt from './DrawerExt';
+import { MixinRangeNumberExt as RangeNumberExt, MixinRangeNumberOutLineExt as RangeNumberOutLineExt } from './RangeNumberExt';
 export * from 'antd';
 import zh_CN from 'antd/locale/zh_CN';
+import en_US from 'antd/locale/en_US';
 import type { Theme } from './ConfigProviderExt/context';
 
 export {
@@ -17,10 +19,13 @@ export {
   SelectExt,
   SelectOutLineExt,
   DrawerExt,
+  RangeNumberExt,
+  RangeNumberOutLineExt,
 
   Theme,
 }
 
 export const locale = {
   zh_CN,
+  en_US,
 }