+++ packages/antd-ext/components/utils/index.ts
@@ -1,0 +1,7 @@
+import useMapTheme from './useMapTheme';
+import usePrefixCls from './usePrefixCls';
+
+export {
+  useMapTheme,
+  usePrefixCls,
+}